# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The Road Of Code",
        "title": "Road",
        "contact": {},
        "license": {
            "name": "Raja",
            "url": "https://github.com/0RAJA"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/star/{post_id}": {
            "get": {
                "description": "通过post_id获取用户对于某个帖子的点赞情况",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_star"
                ],
                "summary": "获取用户对于某个帖子的点赞情况",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "帖子ID",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true表示已经点过赞，false表示没有",
                        "schema": {
                            "type": "bool"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            },
            "put": {
                "description": "设置用户对于某个帖子的点赞",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_star"
                ],
                "summary": "用户对某个帖子点赞",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "帖子ID",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "通过post_id取消用户对于某个帖子的点赞",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user_star"
                ],
                "summary": "取消用户对于某个帖子的点赞",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "帖子ID",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
...
        "/token/get": {
            "get": {
                "description": "用户将被重定向到github授权，服务器会得到授权后返回用户信息和token,retoken",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "获取token",
                "responses": {
                    "200": {
                        "description": "返回用户信息和token以及retoken",
                        "schema": {
                            "$ref": "#/definitions/logic.GetTokenReply"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/token/refresh": {
            "get": {
                "description": "将过期的token和未过期的retoken换取新的token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "刷新token",
                "responses": {
                    "200": {
                        "description": "返回用户信息和新的token",
                        "schema": {
                            "$ref": "#/definitions/logic.RefreshTokenReply"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
...
        "/user/createTime": {
            "get": {
                "description": "通过时间段来检索用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "通过时间段来检索用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码 default 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量 default and max 10",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回帖子简介的数组和描述数组大小的信息",
                        "schema": {
                            "$ref": "#/definitions/logic.ListUsersReply"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/user/users": {
            "get": {
                "description": "列出用户的用户名，头像链接，仓库链接，ip地址，创建时间，修改时间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "列出用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户信息",
                        "schema": {
                            "$ref": "#/definitions/logic.ListUsersReply"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/user/{username}": {
            "get": {
                "description": "获取用户的用户名，头像链接，仓库链接，ip地址，创建时间，修改时间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户信息",
                        "schema": {
                            "$ref": "#/definitions/logic.UserInfo"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/views/all": {
            "get": {
                "description": "获取时间段内的访问数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "views"
                ],
                "summary": "获取时间段内的访问数量",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码 default 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量 default and max 10",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "指定时间内的所有访问数",
                        "schema": {
                            "$ref": "#/definitions/logic.ListViewsByCreateTimeReply"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        },
        "/views/post/{post_id}": {
            "get": {
                "description": "获取指定帖子的新增访问量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "views"
                ],
                "summary": "获取指定帖子的新增访问量",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "帖子ID",
                        "name": "post_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回对应帖子的新增访问量",
                        "schema": {
                            "type": "int64"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errcode.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "logic.GetTokenReply": {
            "type": "object",
            "properties": {
                "re_token": {
                    "description": "ReToken",
                    "$ref": "#/definitions/logic.ReToken"
                },
                "token": {
                    "description": "Token",
                    "$ref": "#/definitions/logic.Token"
                },
                "user": {
                    "description": "用户信息",
                    "$ref": "#/definitions/logic.UserInfo"
                }
            }
        },
...
        "logic.ListUsersReply": {
            "type": "object",
            "properties": {
                "pager": {
                    "$ref": "#/definitions/logic.Pager"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/logic.User"
                    }
                }
            }
        },
        "logic.ListViewsByCreateTimeReply": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/logic.View"
                    }
                },
                "pager": {
                    "$ref": "#/definitions/logic.Pager"
                }
            }
        },
...
        "logic.Pager": {
            "type": "object",
            "properties": {
                "page": {
                    "description": "页数",
                    "type": "integer"
                },
                "page_size": {
                    "description": "一页的大小",
                    "type": "integer"
                },
                "total_rows": {
                    "description": "结果的总数量",
                    "type": "integer"
                }
            }
        },
...
        "logic.ReToken": {
            "type": "object",
            "properties": {
                "expiredAt": {
                    "description": "\"过期时间\"`",
                    "type": "string"
                },
                "refreshToken": {
                    "description": "\"刷新token\"`",
                    "type": "string"
                }
            }
        },
        "logic.RefreshTokenReply": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "Token",
                    "$ref": "#/definitions/logic.Token"
                },
                "user": {
                    "description": "用户信息",
                    "$ref": "#/definitions/logic.UserInfo"
                }
            }
        },
...
        "logic.Token": {
            "type": "object",
            "properties": {
                "expiredAt": {
                    "description": "\"过期时间\"`",
                    "type": "string"
                },
                "token": {
                    "description": "\"token\"`",
                    "type": "string"
                }
            }
        },
        "logic.User": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "IP地址",
                    "type": "string"
                },
                "avatar_url": {
                    "description": "头像链接",
                    "type": "string"
                },
                "create_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "depository_url": {
                    "description": "仓库连接",
                    "type": "string"
                },
                "modify_time": {
                    "description": "修改时间",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "logic.UserInfo": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "description": "\"头像链接\"`",
                    "type": "string"
                },
                "depository_url": {
                    "description": "\"仓库链接\"`",
                    "type": "string"
                },
                "username": {
                    "description": "\"用户名\"`",
                    "type": "string"
                }
            }
        },
        "logic.View": {
            "type": "object",
            "properties": {
                "create_time": {
                    "description": "截止时间",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "views_num": {
                    "description": "访问数",
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}